Implemented search for all pages requiring search and implemented session too

+++ pages/delete_from_cart.php
@@ -4,7 +4,7 @@ include("../classes/product_management.php");
         ini_set('display_errors', 1);
         ini_set('display_startup_errors', 1);
         error_reporting(E_ALL);
-        if(isset($_GET['id'])){
+        if(isset($_SESSION['is_admin']) && $_SESSION['is_admin']==0 && isset($_GET['id'])){
                 $product=new ProductManager();
                 $product->deleteProductFromCart($_GET['id']);
                 header("location:cart_page.php?message='Product deleted successfully'");
